requirements.txt
# Discord関連
discord.py>=2.3.0

# AI API関連
openai>=1.0.0
aiohttp>=3.8.0

# 環境変数管理
python-dotenv>=0.19.0

# その他
requests>=2.28.0
